+++ NEWS.md
@@ -1,3 +1,26 @@
+# microinverterdata 0.2.0
+
+* Polish description and README
+* Use a `switch()` based dispatch method on `model`
+
+# microinverterdata 0.1.4
+
+* Generalize Enphase Envoy support (`get_output_data()` only) #7
+
+# microinverterdata 0.1.3
+
+* Add support to Enphase Energy inverters (`get_output_data()` only) #6
+
+# microinverterdata 0.1.2
+
+* Add mocked http data for APSystems and Fronius #4
+* Add support for 'Fronius' inverters 
+
+# microinverterdata 0.1.1
+
+* Add support to Enphase Envoy-S inverters (`get_output_data()` only) #1
+* manage device unreachable error
+
 # microinverterdata 0.1.0
 
 * Initial CRAN submission.